重构领域对象，定义valid指责，和注解指责。valid将是注解的超集，注解只用于insert验证对象合法性

+++ src/main/java/myblog/service/CategoryService.java
@@ -6,6 +6,7 @@ import myblog.domain.Category;
 import myblog.domain.Pagination;
 import myblog.domain.Sort;
 
+import javax.ws.rs.NotFoundException;
 import java.util.HashMap;
 import java.util.List;
 
@@ -20,6 +21,22 @@ public class CategoryService {
         CategoryDaoMyBatisImpl myBatisCategoryDao = (CategoryDaoMyBatisImpl)
                 DaoFactory.getDaoFactory(DaoFactory.DaoBackend.MYBATIS).getCategoryDao();
 
+        if (category.getCategory_parent_id() != null) {
+            Category parent = myBatisCategoryDao.getCategoryById(category.getCategory_parent_id());
+            if (parent != null) {
+                category.setCategory_root_id(parent.getCategory_root_id());
+                category.setCategory_level(parent.getCategory_level());
+            } else {
+                throw new NotFoundException("Not found parent category: Id = " + category.getCategory_parent_id());
+            }
+        } else {
+            category.setCategory_level(1);
+        }
+
+        category.setCategory_enabled(true);
+        category.setCategory_created_at(null);
+        category.setCategory_updated_at(null);
+
         return myBatisCategoryDao.createCategory(category);
     }
 
@@ -32,7 +49,11 @@ public class CategoryService {
         CategoryDaoMyBatisImpl myBatisCategoryDao = (CategoryDaoMyBatisImpl)
                 DaoFactory.getDaoFactory(DaoFactory.DaoBackend.MYBATIS).getCategoryDao();
 
-        return myBatisCategoryDao.deleteCategory(categoryId);
+        if (myBatisCategoryDao.getCategoryById(categoryId) != null) {
+            return myBatisCategoryDao.deleteCategory(categoryId);
+        } else {
+            throw new NotFoundException("Not found category: Id = " + categoryId);
+        }
     }
 
     /**
@@ -40,11 +61,18 @@ public class CategoryService {
      * @param category
      * @return
      */
-    public static boolean updateCategory(Category category) {
+    public static boolean updateCategory(int categoryId, Category category)  {
         CategoryDaoMyBatisImpl myBatisCategoryDao = (CategoryDaoMyBatisImpl)
                 DaoFactory.getDaoFactory(DaoFactory.DaoBackend.MYBATIS).getCategoryDao();
 
-        return myBatisCategoryDao.updateCategory(category);
+        if (myBatisCategoryDao.getCategoryById(categoryId) != null) {
+            category.setCategory_id(categoryId);
+            category.setCategory_updated_at(null);
+
+            return myBatisCategoryDao.updateCategory(category);
+        } else {
+            throw new NotFoundException("Not found category: Id = " + categoryId);
+        }
     }
 
     /**
@@ -67,6 +95,7 @@ public class CategoryService {
     public static List<Category> getCategories(Category category) {
         CategoryDaoMyBatisImpl myBatisCategoryDao = (CategoryDaoMyBatisImpl)
                 DaoFactory.getDaoFactory(DaoFactory.DaoBackend.MYBATIS).getCategoryDao();
+
         HashMap<String, Object> params = category.convertToHashMap();
 
         return myBatisCategoryDao.getCategoriesByCondition(params);
@@ -82,6 +111,7 @@ public class CategoryService {
     public static Pagination<Category> getCategoryList(Category category, Pagination<Category> page, Sort order) {
         CategoryDaoMyBatisImpl myBatisCategoryDao = (CategoryDaoMyBatisImpl)
                 DaoFactory.getDaoFactory(DaoFactory.DaoBackend.MYBATIS).getCategoryDao();
+
         HashMap<String, Object> params = category.convertToHashMap();
         params.put("limit", page.getLimit());
         params.put("offset", page.getOffset());
